refactor(VR): fold DlssEnhancer into Upscaling, mirror DLSSperf pattern
Following the precedent set in PR-2 for DLSSperf: DlssEnhancer is the
second VR-DLSS surface to live entirely inside Upscaling rather than as
a peer Feature. Both share the same "mode of Upscaling, not its own
Feature" architecture.

Structural changes:
- struct DlssEnhancerFeature : Feature  →  struct DlssEnhancer (drop
  Feature inheritance and the GetName/IsCore/GetCategory boilerplate).
- src/Features/DlssEnhancerFeature.{cpp,h}  →
  src/Features/Upscaling/DlssEnhancer.{cpp,h}
- src/Features/DlssEnhancer/*  →  src/Features/Upscaling/DlssEnhancer/*
- features/DLSS Enhancer/Shaders/Features/SubrectStretchCS.hlsl  →
  features/Upscaling/Shaders/Upscaling/DlssEnhancer/SubrectStretchCS.hlsl
- DLSSENHANCER.ini deleted (placeholder version marker; DLSSperf has no
  .ini either).
- Implementation namespace DlssEnhancer  →  DlssEnhancerImpl to avoid
  collision with the new struct name.

Membership / lifecycle:
- globals::features::dlssEnhancer is gone; access via
  globals::features::upscaling.dlssEnhancer (static-inline member of
  Upscaling, sibling to streamline / fidelityFX / rcas / dlssPerf).
- Removed from globals::features registration and Feature::ms_features.
- PostPostLoad, ClearShaderCache, SaveSettings, LoadSettings,
  RestoreDefaultSettings driven from the equivalent Upscaling methods.

Settings dedup:
- DlssEnhancer::Settings drops qualityMode / sharpnessDLSS / presetDLSS
  / streamlineLogLevel (already on Upscaling::Settings — double-stored).
- DlssEnhancer accessors / Bridge.cpp / Postprocess.cpp now read those
  four through upscaling.settings.

UI:
- DlssEnhancer's standalone panel sliders for the dedup'd fields are
  gone; the Upscaling panel covers them for both DLSS paths.
- DlssEnhancer::DrawSettings renders only DlssEnhancer-specific knobs
  (Enable, DLSS Mode, Stretch Mode, Sharpen Mode, MV/Reactive/
  Transparency masks, Subrect editor).
- Called from Upscaling::DrawSettings inside a "Foveated DLSS
  (DlssEnhancer)" TreeNode, gated on isVR.
- "RESTART REQUIRED" warning replaced with Util::Text::RestartNeeded to
  match the cue convention from PR-2.

JSON:
- DlssEnhancer round-trips under "dlssEnhancer" sub-key in Upscaling's
  settings JSON. Upscaling::LoadSettings extracts and erases the
  sub-block before deserializing its own fields, so absence is safe.
