--- features/DLSS Enhancer/Shaders/Features/DLSSENHANCER.ini
@@ -1,2 +1,0 @@
-[Info]
-Version = 1-0-0
