+++ src/Feature.cpp
@@ -3,7 +3,6 @@
 #include "FeatureIssues.h"
 #include "FeatureVersions.h"
 #include "Features/CloudShadows.h"
-#include "Features/DlssEnhancerFeature.h"
 #include "Features/DynamicCubemaps.h"
 #include "Features/ExponentialHeightFog.h"
 #include "Features/ExtendedMaterials.h"
@@ -241,7 +240,6 @@ const std::vector<Feature*>& Feature::GetFeatureList()
 		&globals::features::ibl,
 		&globals::features::extendedTranslucency,
 		&globals::features::upscaling,
-		&globals::features::dlssEnhancer,
 		&globals::features::renderDoc,
 		&globals::features::remoteControl,
 		&globals::features::weatherEditor,
