+++ src/Features/Upscaling/DlssEnhancer/Preprocess.h
@@ -2,7 +2,7 @@
 
 struct Upscaling;
 
-namespace DlssEnhancer
+namespace DlssEnhancerImpl
 {
 	class Preprocess
 	{
