+++ src/Globals.cpp
@@ -2,7 +2,6 @@
 
 #include "Deferred.h"
 #include "Features/CloudShadows.h"
-#include "Features/DlssEnhancerFeature.h"
 #include "Features/DynamicCubemaps.h"
 #include "Features/ExponentialHeightFog.h"
 #include "Features/ExtendedMaterials.h"
@@ -86,7 +85,6 @@ namespace globals
 		ExtendedTranslucency extendedTranslucency{};
 		Upscaling upscaling{};
 		HDRDisplay hdrDisplay{};
-		DlssEnhancerFeature dlssEnhancer{};
 		RenderDoc renderDoc{};
 		RemoteControl remoteControl{};
 		ScreenshotFeature screenshotFeature{};
